+++ common/src/main/java/net/picopress/mc/mods/zombietactics2/mixin/MonsterMixin.java
@@ -5,7 +5,7 @@ import net.picopress.mc.mods.zombietactics2.Config;
 import net.minecraft.core.BlockPos;
 import net.minecraft.util.RandomSource;
 import net.minecraft.world.Difficulty;
-import net.minecraft.world.entity.monster.Zombie;
+import net.minecraft.world.entity.monster.zombie.Zombie;
 import net.minecraft.world.entity.EntityType;
 import net.minecraft.world.entity.PathfinderMob;
 import net.minecraft.world.entity.monster.Monster;