+++ common/src/main/java/net/picopress/mc/mods/zombietactics2/mixin/ZombieMisc.java
@@ -1,6 +1,6 @@
 package net.picopress.mc.mods.zombietactics2.mixin;
 
-import net.minecraft.world.entity.monster.Zombie;
+import net.minecraft.world.entity.monster.zombie.Zombie;
 
 import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.gen.Accessor;