+++ common/src/main/java/net/picopress/mc/mods/zombietactics2/mixin/ZombieVillagerMixin.java
@@ -3,8 +3,7 @@ package net.picopress.mc.mods.zombietactics2.mixin;
 import net.picopress.mc.mods.zombietactics2.Config;
 
 import net.minecraft.world.entity.EntityType;
-import net.minecraft.world.entity.monster.Zombie;
-import net.minecraft.world.entity.monster.ZombieVillager;
+import net.minecraft.world.entity.monster.zombie.*;
 import net.minecraft.world.level.Level;
 
 import org.spongepowered.asm.mixin.Mixin;